+++ R/zzz.R
@@ -3,22 +3,22 @@
 . = NULL # don't look unbound
 
 
+#' Set rqdatatable package as default rquery executor
+#'
+#' Sets rqdatatable (and hence data.table) as the default executor for rquery).
+#'
+#' @export
+#'
+set_rqdatatable_as_executor <- function() {
+  options(list("rquery.rquery_executor" = list(f = ex_data_table, name = "rqdatable")))
+  invisible(NULL)
+}
+
 .onAttach <- function(libname, pkgname) {
+  # attach happens after load, so set as the executor only in the attach case
   prev_exec <- getOption("rquery.rquery_executor", default = NULL)
   if(is.null(prev_exec) || (is.list(prev_exec) && isTRUE(prev_exec$name == "rqdatable"))) {
-    options(list("rquery.rquery_executor" = list(f = ex_data_table, name = "rqdatable")))
-  } else {
-    packageStartupMessage("rqdatatable loaded, but did not register itself as the executor (already one registered)")
+    set_rqdatatable_as_executor()
   }
-  invisible()
+  invisible(NULL)
 }
-
-
-.onLoad <- function(libname, pkgname) {
-  prev_exec <- getOption("rquery.rquery_executor", default = NULL)
-  if(is.null(prev_exec) || (is.list(prev_exec) && isTRUE(prev_exec$name == "rqdatable"))) {
-    options(list("rquery.rquery_executor" = list(f = ex_data_table, name = "rqdatable")))
-  }
-  invisible()
-}
-